requirements.txt
-i https://pypi.org/simple
asgiref==3.2.7
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
django==3.0
djangorestframework==3.11.0
idna==2.9
pytz==2020.1
requests==2.23.0
soupsieve==2.0.1
sqlparse==0.3.1
urllib3==1.25.9
